media: i2c: imx477: Allow control of on-sensor DPC
commit  599c995b61c3e9b3047c4cc74108736263bf002f from
https://github.com/raspberrypi/linux.git rpi-5.15.y

A module parameter "dpc_enable" is added to allow the control of the
sensor's on-board DPC (Defective Pixel Correction) function.

This is a global setting to be configured before using the sensor;
there is no intention that this would ever be changed on-the-fly.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>