media: i2c: imx477: Allow control of on-sensor DPC
commit  0539aa84e4b42ca22d48dc6a003425cd5c1f4f17 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

A module parameter "dpc_enable" is added to allow the control of the
sensor's on-board DPC (Defective Pixel Correction) function.

This is a global setting to be configured before using the sensor;
there is no intention that this would ever be changed on-the-fly.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>