media: i2c: imx477: Allow control of on-sensor DPC
commit  a9253caafe33cfd401395215bb9f2679cbb02dcf from
https://github.com/raspberrypi/linux.git rpi-5.16.y

A module parameter "dpc_enable" is added to allow the control of the
sensor's on-board DPC (Defective Pixel Correction) function.

This is a global setting to be configured before using the sensor;
there is no intention that this would ever be changed on-the-fly.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>