regulator: rpi-panel: Serialise operations.
commit  50e5c1a2b7fb2afd2580eea013a5cb5730cd12c8 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

The driver was using the regmap lock to serialise the
individual accesses, but we really need to protect the
timings of enabling the regulators, including any communication
with the Atmel.

Use a mutex within the driver to control overall accesses to
the Atmel, instead of the regmap lock.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>