regulator: rpi-panel: Serialise operations.
commit  210ed93a37557c2b661e5f1f15a2b0edc6af76fa from
https://github.com/raspberrypi/linux.git rpi-5.16.y

The driver was using the regmap lock to serialise the
individual accesses, but we really need to protect the
timings of enabling the regulators, including any communication
with the Atmel.

Use a mutex within the driver to control overall accesses to
the Atmel, instead of the regmap lock.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>