drm/vc4: Reset HDMI MISC_CONTROL register.
commit  2476504dccce03ea375738a4ecbca9d42550b9ee from
https://github.com/raspberrypi/linux.git rpi-5.16.y

The HDMI block can repeat pixels for double clocked modes,
and the firmware is now configuring the block to do this as
the PV is doing it incorrectly when at 2pixels/clock.
If the kernel doesn't reset it then we end up with strange
modes.

Reset MISC_CONTROL.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>