regulator: rpi-panel: Remove get_brightness hook
commit  079563b852f9446d72f09b87bdcdcf5c3f948a33 from
https://github.com/raspberrypi/linux.git rpi-5.16.y

The driver was implementing a get_brightness function that
tried to read back the PWM setting of the display to report
as the current brightness.
The controller on the display does not support that, therefore
we end up reporting a brightness of 0, and that confuses
systemd's backlight service.

Remove the hook so that the framework returns the current
brightness automatically.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>