regulator: rpi-panel: Remove get_brightness hook
commit  942f7ee98576df14e5b4d2edc01240199b49bfa4 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

The driver was implementing a get_brightness function that
tried to read back the PWM setting of the display to report
as the current brightness.
The controller on the display does not support that, therefore
we end up reporting a brightness of 0, and that confuses
systemd's backlight service.

Remove the hook so that the framework returns the current
brightness automatically.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>