drm/vc4: hdmi: Simplify the hotplug handling
commit  380f2eb4cf85c3e3def9192701a309b2cebec513 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

Our detect callback has a bunch of operations to perform depending on
the current and last status of the connector, such a setting the CEC
physical address or enabling the scrambling again.

This is currently dealt with a bunch of if / else statetements that make
it fairly difficult to read and extend.

Let's move all that logic to a function of its own.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>