dtoverlays: Drop i2c baudrate for Pi 7inch DSI screen to 50kHz
commit  d1bbf885275f926b90b43c4905f171fe612308e4 from
https://github.com/raspberrypi/linux.git rpi-5.16.y

Whilst investigations continue as to why the 7" DSI screen doesn't
work on Pi3 with KMS, drop the I2C baudrate to 50kHz as that seems
to be a reliable workaround.
This will be reverted once the full issue is understood.

https://github.com/raspberrypi/linux/issues/4686

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>