staging/bcm2835-codec: Allow a different stride alignment per role
commit  50f7a32473f41e2075ca13726fe580b9524446c1 from
https://github.com/raspberrypi/linux.git rpi-5.16.y

Deinterlace and decode aren't affected in the same way as encode
and ISP by the alignment requirement on 3 plane YUV420.
Decode would be affected, but it always aligns the height up to
a macroblock, and uses the selection API to reflect that.

Add in the facility to set the bytesperline alignment per role.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>