drm/v3d: Allow -EACCES return from pm_runtime_get*
commit  8aea3e65552faed532beac961e26c08ccdab1bc4 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

We are currently disabling PM in v3d because it causes a variety of
errors:

[   13.521599] v3d fec00000.v3d: TLB clear wait idle pre-wait failed
[   13.621884] v3d fec00000.v3d: TLB clear wait idle failed
[   13.621906] v3d fec00000.v3d: MMU flush timeout

pm_runtime_get_sync and friends can return -EACCES if disable_depth is
non-zero, indicating that runtime PM is disabled (or not yet enabled).
This is not an error, and can be ignored (although we should probably
get to the bottom of the errors and re-enable it at some point).

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>