drm/vc4: hdmi: Always try to have the highest bpc
commit  6b45f2b1aad45d4aa4a6d3ee8877b2013ef4e585 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

Currently we take the max_bpc property as the bpc value and do not try
anything else.

However, what the other drivers seem to be doing is that they would try
with the highest bpc allowed by the max_bpc property and the hardware
capabilities, test if it results in an acceptable configuration, and if
not decrease the bpc and try again.

Let's use the same logic.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>