drm/vc4: hdmi: Take the sink maximum TMDS clock into account
commit  abfb521e5157f508d355147d0d7691f345003618 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

In the function that validates that the clock isn't too high, we've only
taken our controller limitations into account so far.

However, the sink can have a limit on the maximum TMDS clock it can deal
with too which is exposed through the EDID and the drm_display_info.

Make sure we check it.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>