drm/vc4: Convert vc4_dsi to using a bridge instead of encoder.
commit  8f8cd644577d3b378ce583e9ec6e18e4aab3984b from
https://github.com/raspberrypi/linux.git rpi-6.0.y

Remove the encoder functions, and create a bridge attached to
this dumb encoder which implements the same functionality.

As a bridge has state which an encoder doesn't, we need to
add the state management functions as well.

As there is no bridge atomic_mode_set, move the initialisation
code that was in mode_set into _pre_enable.
The code to actually enable and disable sending video are split
from the general control into _enable and _disable.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>