media: i2c: ov9281: Increase diff between VTS and max exposure
commit  4d209208f4ef0e2f29222ba8691c0186d4c8e2a3 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

The driver did allow the exposure to go up to VTS - 4 lines,
but this would produce a visible line on 1280x800, and a stall of
the sensor at 640x480.

Whilst it appears to work with a difference of 5, the datasheet states
there should be at least 25 lines difference between VTS and exposure,
so use that value.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>