pwm: raspberrypi-poe: Add option of being created by MFD or FW
commit  751280cff052d89db1889e451c291aebf12806d1 from
https://github.com/raspberrypi/linux.git rpi-6.2.y

The firmware can only use I2C0 if the kernel isn't, therefore
with libcamera and DRM using it the PoE HAT fan control needs
to move to the kernel.

Add the option for the driver to be created by the PoE HAT core
MFD driver, and use the I2C regmap that provides to control fan
functions.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>