usb: xhci: add a quirk for Superspeed bulk OUT transfers on VL805
commit  aa8a74bd2d59026901a04c07e6e3bffa97e77b3c from
https://github.com/raspberrypi/linux.git rpi-6.1.y

The VL805 has a bug in its internal FIFO space accounting that results
in bulk OUT babble if a TRB in a large multi-element TD has a data
buffer size that is larger than and not a multiple of wMaxPacketSize for
the endpoint. If the downstream USB3.0 link is congested, or latency is
increased through an intermediate hub, then the VL805 enters a suspected
FIFO overflow condition and transmits repeated, garbled data to the
endpoint.

TDs with TRBs of exact multiples of wMaxPacketSize and smaller than
wMaxPacketSize appear to be handled correctly, so split buffers at a
1024-byte length boundary and put the remainder in a separate smaller
TRB.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>