staging/vc-sm-cma: Avoid log spamming on Pi0/1 over cache alias.
commit  44e5802ba3c8ebe9aa8634c518753ee9c3e56c56 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

Pi 0/1 use the 0x80000000 cache alias as the ARM also sees the world
through the VPU L2 cache.
vc-sm-cma was trying to ensure it was in an uncached alias (0xc), and
complaining on every allocation if it weren't. Reduce this logging.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>