media: media.h: Add new media link type
commit  1a8cc72706e08c91c10119cb6a42a03cf9951757 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

To describe in the kernel the connection between devices and their
supporting peripherals (for example, a camera sensor and the vcm
driving the focusing lens for it), add a new type of media link
to introduce the concept of these ancillary links.

Add some elements to the uAPI documentation to explain the new link
type, their purpose and some aspects of their current implementation.

Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Daniel Scally <djrscally@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>