drm/vc4: txp: Force alpha to be 0xff if it's disabled
commit  2bed5b1ed85185207b0520c14a14e644ab7b98fc from
https://github.com/raspberrypi/linux.git rpi-5.18.y

If we use a format that has padding instead of the alpha component (such
as XRGB8888), it appears that the Transposer will fill the padding to 0,
disregarding what was stored in the input buffer padding.

This leads to issues with IGT, since it will set the padding to 0xff,
but will then compare the CRC of the two frames which will thus fail.

Fixes: 008095e065a8 ("drm/vc4: Add support for the transposer block")
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>