drm/vc4: hdmi: Add CSC for BT601/709/2020 limited and full range output
commit  470743b630997f2e35732ff6f1082af553a196e0 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

The HVS always composes in the RGB domain, but there is a colourspace
conversion block on the output to allow for sending YCbCr over the
HDMI interface.
The colourspace on that link is configurable via the "Colorspace"
property on the connector, and that updates the infoframes. There
is also selection of limited or full range based on the mode selected
or an override.

Add code to update the CSC as well so that the metadata matches the
image data.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>