media: i2c: Rename ad5398 to ad5398_vcm
commit  ccfb2f1da2d4dd177aa30a62580aa0c492fe4d3f from
https://github.com/raspberrypi/linux.git rpi-6.0.y

There's already a regulator module called ad5398 that exposes
this device through the regulator API. That is meaningless in
the terms that it uses and how it maps to V4L2, so a new driver
was added. However the module name collision wasn't noted, so
rename it now.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>