media: i2c: Rename ad5398 to ad5398_vcm
commit  dae74ec94a7879c1ba7a181da38d79c0576621a2 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

There's already a regulator module called ad5398 that exposes
this device through the regulator API. That is meaningless in
the terms that it uses and how it maps to V4L2, so a new driver
was added. However the module name collision wasn't noted, so
rename it now.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>