media: i2c: Rename ad5398 to ad5398_vcm
commit  f9b632448de94c7e6d9784bdf10c5c310b0dc744 from
https://github.com/raspberrypi/linux.git rpi-6.2.y

There's already a regulator module called ad5398 that exposes
this device through the regulator API. That is meaningless in
the terms that it uses and how it maps to V4L2, so a new driver
was added. However the module name collision wasn't noted, so
rename it now.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>