drm/vc4: Make sure we don't end up with a core clock too high
commit  3a458f569ff3d29aed80e39c6049e2228282bfc2 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

Following the clock rate range improvements to the clock framework,
trying to set a disjoint range on a clock will now result in an error.

Thus, we can't set a minimum rate higher than the maximum reported by
the firmware, or clk_set_min_rate() will fail.

Thus we need to clamp the rate we are about to ask for to the maximum
rate possible on that clock.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>